Updated goldens.

+++ FinModelUtility/Formats/Glo/Glo Tests/goldens/chuck/output/material2.fragment.glsl
@@ -2,7 +2,7 @@
 
 uniform sampler2D diffuseTexture;
 uniform float shininess;
-uniform float useLighting;
+uniform int useLighting;
 
 out vec4 fragColor;
