+++ FinModelUtility/Formats/Glo/Glo Tests/goldens/joff/output/material4.fragment.glsl
@@ -4,7 +4,7 @@ uniform sampler2D diffuseTexture;
 uniform sampler2D ambientOcclusionTexture;
 uniform sampler2D emissiveTexture;
 uniform float shininess;
-uniform float useLighting;
+uniform int useLighting;
 
 out vec4 fragColor;
