+++ FinModelUtility/Games/HaloWars/HaloWars Tests/goldens/xtd/blood_gulch/output/material0.fragment.glsl
@@ -24,7 +24,7 @@ uniform sampler2D diffuseTexture;
 uniform sampler2D ambientOcclusionTexture;
 uniform sampler2D emissiveTexture;
 uniform float shininess;
-uniform float useLighting;
+uniform int useLighting;
 
 out vec4 fragColor;
